requirements.web.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.8.2
PyYAML>=6.0.2
psutil>=5.9.8
